Make StatisticalInverseProblem use SG base class
I'm fairly certain this is not technically backwards compatible, but I'm also
not entirely convinced our users are working with the QUESO sequence generator
on such a fine-grained level.  Perhaps it's better to add an alternative getter
for the base class and leave the one the returns the derived class in there but
deprecate it.

This also makes the generateSequence() method in the base class pure virtual.
This makes sense because this should be totally determined by the derived
class, with the derived class being being the class that has specific sampler
behaviour.