Add setSequenceGenerator test
The following changes were made to make setting the sequence generator work.
First, we only set the sequence generator to MetropolisHastingsSG if the
sequence generator is NULL and therefore hasn't already been set.  Second,
the sequence generator needs the posterior random variable (actually only the
pdf) but this isn't set in the StatisticalInverseProblem ctor, so we do that.
I imagine this will introduce some memory leaks.