Make the parameter vector *serial* in the lhd call
Update the document to reflect that it's much easier from the user's
perspective to interact with a serial parameter vector that can be
passed off to their parallel forward code.  Also improves compatibility.
Existing parallel forward models wouldn't need to be changed to
understand the way QUESO partiions vectors.  It also means the user
isn't burdened with point-to-point communication to organise the
partitioned vector into the correct form.