Use pointer for stored env, and add comm map
The pointer for the stored env is so we can do run-time polymorphism on
the queso environment; sometimes the environment is EmptyEnvironment and
sometimes it is a FullEnvironment.

I added a static map as per Roy's suggestion so that we can easily
convert between queso communicators and libmesh communicators.  In
particular, when queso constructs a vector object it passes its own
queso-branded version of an mpi communicator.  Since the vector object
we'd like to construct is a GslNumericVector, which is a subclass of
NumericVector, we have to be able to call the NumericVector base class
constructor in the initialisation list of GslNumericVector and the
NumericVector constructor expects a libmesh-branded communicator.  Not a
queso-branded one.  The map provides the mechanism to translate from one
to the other in the initialisation list of GslNumericVector.