Reimplement invertMultiply for GslNumericBlockMatrix
We had to reimplement this (and store our own blocks and vector spaces
to work around the fact that when you copy a block belonging to the
internal type and return a reference to the user, any changes the user
makes would not propagate to the internal type.  So that's why we have
to store our own.

Also fix the fact the dtor was throwing an exception.