Adding dummy impls of cwSet* statistical methods
They require that the vector have access to the environment to query the
concrete Rng object.  The Rng object needs to be constructed with a seed
and an mpi rank.

The cwSet* statistical methods should really be non-member non-friend
functions that take a vector and an environment as arguments instead.
This aligns with Meyers's philosophy, "Prefer non-member non-friend
functions."