Make MiscCheckTheParallelEnv take an BaseEnv
Because pulling it from a vector is not the right approach.  All the
places this function gets called is from a class that already has an env
reference so this is not a big deal.  I just hope it's not used in user
code as this is a backwards-incompatible change.